lasermx/requirements.txt



pyserial>=3.5
PySide6>=6.6
svgpathtools>=1.6.1
ezdxf>=1.3
shapely>=2.0
